requirements.txt
mysql-connector-python==8.0.33
pandas==2.0.3
python-dotenv==1.0.0
openpyxl==3.1.2
jupyter=1.1.1
